Get css imports to work with jest
This configures jest to use the themeable transform for css imports.

Note: this doesn't use css modules but uses identity-obj-proxy.

This is the recommended approach in the jest docs:

https://facebook.github.io/jest/docs/webpack.html#handling-static-assets

If we need to, we could make this actually run postcss and maybe use
something like postcss-modules, but if we don't care what the rendered
markup looks like this should work.

Change-Id: Id817702c61ba79a5f0e5a634eaa0678a6178eb3e
Reviewed-on: https://gerrit.instructure.com/109017
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>