Use container queries instead of media queries
Fixes FALCOR-294

This makes the component more reusable because they will adjust
according the container they are rendered in, intead of the viewport.

Also note that the query names can be whatever makes sense and the values
can be defined according to what works best for the content in the
component.

Change-Id: Iaae96f5f3ba95d82cf78086c7aa9532552e6f12b
Reviewed-on: https://gerrit.instructure.com/111835
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>