make new activity button stick to the header
it used to stick to the top of the window. this makes it stick to the
bottom of the header.

this flickers in chrome, but that's chrome, not us. The button doesn't
stick in IE.

this only affects the demo area for now. canvas will need to calculate
the offset in a separate canvas commit.

fixes FALCOR-366

test plan:
* scroll down and see new activity sticks at the bottom of the header

Change-Id: I8aedbb73c8d304626b794aa9f97b465cc4ab4907
Reviewed-on: https://gerrit.instructure.com/115923
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>