coordinate a11y focus and scrolling animations
When loading new items, we don't want the view to jump suddenly and
disorient the user. We want to visually keep the scroll position the
same. But for a11y we also need to set the keyboard focus to one of the
newly loaded items, which will jump to the element receiving focus.
To meet both of these requirements, we will animate scrolling to the
newly loaded item that is receiving focus.

We accomplish this by having a dynamic ui manager that coordinates the
focus and animation actions that need to happen based on the data that
was loaded. This centralizes the problem instead of spreading the
responsibility out to all the ui components.

The UI components register themselves with the manager and provide the
focusable and scrollable elements the manager can use to perform the
animations. The manager is also informed of the redux actions happening
in the application via redux middleware so it can determine what
animations should run. The top level ui then notifies the manager to
perform those animations at the appropriate time, after all the new ui
elements have been created and mounted.

fixes: FALCOR-370

Change-Id: I84fba94dfb903a207ad03b9ce7db8f6560d6eea9
Reviewed-on: https://gerrit.instructure.com/116746
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>