Cope with nonexistent course in a planner item
The planner/item api used to return notes from concluded courses with a
course_id, but the course detail data was unavailable and loading the
api results failed.  The changes here check that we have course detail
data before trying to add it to the note, and if not just skips it. This
has the effect of converting a course toto to a vanilla todo.

With a companion commit in g/131750, the planner/items api no longer
returns notes from concluded courses, so this should not be an issue,
but the code here is still good defensive programming and will be left
in place

references ADMIN-117

Test plan:
  Once the api is updated, there is no good way to test this.  If you
want to pull this change before g/131750, you will be able to test like
this:
  - create a course, add a student
  - log in as the student and create a planner todo for the course
  - return to the course as teacher|admin and conclude the course
  - return to the planner as the student
  > expect the todo to be in the planner, not associated with any course

Change-Id: I7b78c876177a860a67fb264a63bdfbdc935bd853
Reviewed-on: https://gerrit.instructure.com/131702
Tested-by: Jenkins
Reviewed-by: Mysti Sadler <mysti@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>