Keep Opportunities popup on the screen
Popover (via Position) will reposition the Popoer above the trigger if
there isn't enough room below it, making it unreachable.  Deal with this
by computing the available vertical space, less some buffer, and setting
this as the maximum height on the Popover's content.  This will keep it
below the trigger, and visible. In the event it still doesn't fit, the
Popover is forced to be below and to the left of the trigger.

fixes ADMIN-707

test plan:
  - create a scenario where there are a number of Opportunities, say 4
    missed assignments.
  - shrink the browser so it is only about 500px tall (half a monitor
    height)
  - click the bell button
  > expect the Opportunities popup to be below the bell button and be
  > sized to fit the screen.
  - close the popup
  - shrink the browser to be about 200px tall
  - click the bell button
  > expect the Opportunities popup to be below the bell, but you can
  > scroll it to see all the content.

Change-Id: I61cc0d8f055f8283355c03fd1c220af8da73406d
Reviewed-on: https://gerrit.instructure.com/138713
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
Reviewed-by: Mysti Sadler <mysti@instructure.com>
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>